@@ -4,7 +4,7 @@
   "description": "React components to wrap Velocity animations",
   "main": "index.js",
   "scripts": {
-    "demo": "./node_modules/.bin/webpack-dev-server --progress --colors --content-base demo --hot --inline --config demo/webpack.config.js"
+    "demo": "webpack-dev-server --progress --colors --content-base demo --hot --inline --config demo/webpack.config.js"
   },
   "repository": {
     "type": "git",
@@ -22,29 +22,31 @@
   },
   "homepage": "https://github.com/twitter-fabric/velocity-react#readme",
   "dependencies": {
+    "create-react-class": "^15.5.2",
     "lodash": "^3.10.1",
-    "react-addons-transition-group": "^0.14.0 || ^15.0.0",
-    "velocity-animate": "^1.4.0"
+    "prop-types": "^15.5.8",
+    "react-addons-transition-group": "^15.5.2",
+    "velocity-animate": "^1.5.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0",
-    "react-dom": "^0.14.0 || ^15.0.0"
+    "react": "^0.14.9 || ^15.5.4",
+    "react-dom": "^0.14.9 || ^15.5.4"
   },
   "devDependencies": {
-    "babel-core": "^6.21.0",
-    "babel-loader": "^6.2.10",
-    "babel-plugin-transform-object-rest-spread": "^6.20.2",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^6.4.1",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "css-loader": "^0.19.0",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0",
-    "react-hot-loader": "^1.2.9",
-    "react-tween-state": "^0.1.3",
-    "style-loader": "^0.12.3",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-hot-loader": "^1.3.1",
+    "react-tween-state": "^0.1.5",
+    "style-loader": "^0.12.4",
     "twemoji": "^1.4.1",
-    "underscore.string": "^3.2.0",
-    "webpack": "^1.12.0",
-    "webpack-dev-server": "^1.10.1"
+    "underscore.string": "^3.3.4",
+    "webpack": "^1.15.0",
+    "webpack-dev-server": "^1.16.3"
   }
 }
