@@ -1,9 +1,11 @@
 var _ = require('lodash');
 var React = require('react');
+var PropTypes = require('prop-types');
+var createReactClass = require('create-react-class');
 
 require('../css/polyfill.css');
 
-var Box = React.createClass({
+var Box = createReactClass({
   displayName: 'Box',
 
   statics: {
@@ -15,9 +17,9 @@ var Box = React.createClass({
   },
 
   propTypes: {
-    style: React.PropTypes.object,
-    underneath: React.PropTypes.bool,
-    instruction: React.PropTypes.string,
+    style: PropTypes.object,
+    underneath: PropTypes.bool,
+    instruction: PropTypes.string,
   },
 
   getDefaultProps: function () {
