@@ -1,15 +1,17 @@
 var React = require('react');
+var PropTypes = require('prop-types');
+var createReactClass = require('create-react-class');
 var twemoji = require('twemoji');
 var s = require('underscore.string');
 
 // Component that uses the Twitter twemoji library to turn emoji-rich strings into spans with
 // <img> tags in them (for cross-browser/os compatibility).
-var EmojiSpan = React.createClass({
+var EmojiSpan = createReactClass({
   displayName: 'Emoji',
 
   propTypes: {
-    size: React.PropTypes.number,
-    children: React.PropTypes.string,
+    size: PropTypes.number,
+    children: PropTypes.string,
   },
 
   getDefaultProps: function () {
