@@ -44,18 +44,20 @@ var _ = {
   omit: require('lodash/object/omit'),
 };
 var React = require('react');
+var PropTypes = require('prop-types');
+var createReactClass = require('create-react-class');
 var ReactDOM = require('react-dom');
 var Velocity = require('./lib/velocity-animate-shim');
 
-var VelocityComponent = React.createClass({
+var VelocityComponent = createReactClass({
   displayName: 'VelocityComponent',
 
   propTypes: {
-    animation: React.PropTypes.any,
-    children: React.PropTypes.element.isRequired,
-    runOnMount: React.PropTypes.bool,
-    targetQuerySelector: React.PropTypes.string,
-    interruptBehavior: React.PropTypes.string,
+    animation: PropTypes.any,
+    children: PropTypes.element.isRequired,
+    runOnMount: PropTypes.bool,
+    targetQuerySelector: PropTypes.string,
+    interruptBehavior: PropTypes.string,
     // Any additional properties will be sent as options to Velocity
   },
 
@@ -102,10 +104,6 @@ var VelocityComponent = React.createClass({
     config = config || {};
 
     this._shouldRunAnimation = false;
-
-    if (!this.isMounted() || this.props.animation == null) {
-      return;
-    }
 
     if (config.stop) {
       Velocity(this._getDOMTarget(), 'stop', true);
