@@ -1,6 +1,7 @@
 var _ = require('lodash');
 
 var React = require('react');
+var createReactClass = require('create-react-class');
 var VelocityComponent = require('../../velocity-component');
 var VelocityTransitionGroup = require('../../velocity-transition-group');
 
@@ -20,7 +21,7 @@ var CAPITALS = ['Montgomery', 'Juneau', 'Phoenix', 'Little Rock', 'Sacramento', 
   'Oklahoma City', 'Salem', 'Harrisburg', 'Providence', 'Columbia', 'Pierre', 'Nashville', 'Austin', 'Salt Lake City',
   'Montpelier', 'Richmond', 'Olympia', 'Charleston', 'Madison', 'Cheyenne'];
 
-var CrossfadeExample = React.createClass({
+var CrossfadeExample = createReactClass({
   displayName: 'CrossfadeExample',
 
   getInitialState: function () {
